Keyboard: Fractal layout macro and readme cleanup (#4789)
* Fractal: layout macro cleanup

Removed KEYMAP alias for LAYOUT_preonic_mit. Alias was unused, and the terminology usage is deprecated.

* Fractal: readme cleanup

Updated the line containing the documentation links in the readme.

* Fractal: disable preonic_mit community layout

QMK doesn't have a preonic_mit community layout at this time.

Disables preonic_mit as an option for community layout. Can be re-enabled when the layout is added.